Added some extra code on 32 bit path

+++ libThirdParty32.h
@@ -5,11 +5,14 @@
 #ifndef MCV_LIBTHIRDPARTY32_H
 #define MCV_LIBTHIRDPARTY32_H
 
+#include <cstdint>
+
 // Imagine this is linked from a third party lib
 class ThirdPartyLib {
 public:
     int calculate( int data ) {
-        return data*2;
+//        return (data+1)/2;
+        return (uint16_t)data*2;
     }
 };
